apps/voice/requirements.txt
vocode==0.1.114
deepgram-sdk>=3.0.0
elevenlabs>=1.0.0
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
pydantic>=2.0.0
redis>=5.0.0
twilio>=8.10.0
